[EdgeTPU] Update the ToolchainProvider.test.ts file
This commit updates the ToolchainProvider.test.ts file so that the testing codes work independent from addition of ToolchainEnv.

AS-IS:
Even though the purpose is not to test the registration of backend and toolchainEnv, the test codes should be updated whenever new backends and toolchainEnvs are registered.

TO-BE:
The test codes only use the toolchainEnv of ONE and dummy_backend regardless of backend and toolchainEnv registered in the extension.ts file.

ONE-vscode-DCO-1.0-Signed-off-by: Bumsoo Ko <rhqjatn2398@naver.com>