[EdgeTPU] Update the Install Quick Input test code
This commit updates the InstallQuickInput.test.ts file so that the testing codes work independent from addition of Backend and ToolchainEnv.

AS-IS:
Even though the purpose is not to test the registration of backend and toolchainEnv, the test codes should be updated whenever new backends and toolchainEnvs are registered.

TO-BE:
The test codes only use the Backend and toolchainEnv of ONE and mockup regardless of backend and toolchainEnv registered in the extension.ts file.

ONE-vscode-DCO-1.0-Signed-off-by: profornnan <profornnan@naver.com>